This commit is a massive refactor. Starting at the database, I fixed a major pain point with Matchup linking.  The table structure now has a proper cascade.  I removed the ON DELETE CASCADES, moving that logic into the Provider. NOTE: I did not drop or change the in place tables.  I created new tables (named <Table>V2).  I did not want to break any code branches that were still working.
I refactored a lot of the database code to minimize the number of methods in the Interface, and I moved a lot of the work into memory.

I refactored some of the view code to push business work into the business layer.  I also refactored a bunch of the spots where we were making multiple DB calls to get resources, instead referencing instances from memory.

I redesigned the TournamentViewUI a bit.  I added a new Grid for the score entry form, which gives us the ability to easily toggle visibility on that form.  I added this toggle functionality.

I restored the icons on the matchup list.  They now update properly.

I restored the functionality of the unfinished matchups checkbox.

I think that's everything.  It's late, so I may have missed something.